Renamed KeyboardTest.py to program-UnitTests.py
program-UnitTests.py will now have all the unit tests to test the
various functions that need extensive testing
